refactor(modules): move UI/logging ports out of executor
Relocate action_log, highlight, and runtime_vars into sqyre-ui-model,
domain-coupled port traits into sqyre-ports, and log drawing helpers
into executor log_draw; update imports so all targeted crates compile
and test cleanly after rebasing onto fix/review-findings.
