+++ crates/sqyre-executor/Cargo.toml
@@ -14,6 +14,7 @@ image = { workspace = true }
 parking_lot = { workspace = true }
 rayon = { workspace = true }
 sqyre-domain = { path = "../sqyre-domain" }
+sqyre-ui-model = { path = "../sqyre-ui-model" }
 sqyre-match = { path = "../sqyre-match" }
 sqyre-vision = { path = "../sqyre-vision" }
 sqyre-hotkeys = { path = "../sqyre-hotkeys" }
