+++ crates/sqyre-executor/src/actions/submacro.rs
@@ -1,7 +1,7 @@
 //! Nested macro orchestration: RunMacro.
 
 use crate::error::{ExecError, Result};
-use crate::highlight::{highlight_clear, highlight_cursor, highlight_fill};
+use sqyre_ui_model::{highlight_clear, highlight_cursor, highlight_fill};
 use crate::run::{execute_action, Executor};
 use sqyre_domain::{ActionId, ActionKind, Macro};
