+++ crates/sqyre-ports/Cargo.toml
@@ -11,4 +11,5 @@ workspace = true
 [dependencies]
 image = { workspace = true }
 rayon = { workspace = true }
+sqyre-domain = { path = "../sqyre-domain" }
 thiserror = { workspace = true }
