+++ crates/sqyre-ui-model/Cargo.toml
@@ -11,4 +11,5 @@ workspace = true
 [dependencies]
 parking_lot = { workspace = true }
 sqyre-domain = { path = "../sqyre-domain" }
+sqyre-match = { path = "../sqyre-match" }
 sqyre-varref = { path = "../sqyre-varref" }
