+++ crates/sqyre-ui-model/src/highlight.rs
